+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 #### BREAKING CHANGES:
 
+* PAPI
+    * Added validation to raise an error if the creation of the `akamai_edge_hostname` resource is attempted with an existing edge hostname.
+    * Added validation to raise an error during the update of `akamai_edge_hostname` resource for the fields 'product_id' and 'certificate'.